backend/pyproject.toml
[project]
name = "notme"
version = "0.1.0"
description = "NotMe is a service that provides an anonymous, secure URL for sharing files or text — designed for one-time access and complete privacy."
authors = [
    {name = "Pankaj vishw",email = "vishw.dev.1000@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "django-split-settings (>=1.3.2,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
